requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.0.1
et_xmlfile==2.0.0
Flask==2.0.1
Flask-Login==0.5.0
Flask-MySQLdb==1.0.1
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.2.0
MarkupSafe==2.0.1
mysqlclient==2.0.3
nltk==3.8.1
numpy==2.1.3
openpyxl==3.1.2
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
PyMySQL==1.0.2
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2023.3
regex==2024.11.6
requests==2.28.2
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.5.0
six==1.16.0
sklearn-features==0.0.2
threadpoolctl==3.1.0
tqdm==4.65.0
tzdata==2024.2
urllib3==1.26.15
Werkzeug==2.0.1
xlrd==2.0.1
XlsxWriter==3.0.3
